pyproject.toml
[project]
name = "sakhi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "replicate>=1.0.4",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==3.7.1",
    "asttokens==3.0.0",
    "async-timeout==5.0.1",
    "asyncpg==0.31.0",
    "attrs==25.4.0",
    "av==16.1.0",
    "backoff==2.2.1",
    "backports-asyncio-runner==1.2.0",
    "bcrypt==5.0.0",
    "boto3>=1.35.0",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "cftime==1.6.4.post1",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "dask==2025.9.1",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docker==7.1.0",
    "docstring-parser==0.17.0",
    "email-validator==2.3.0",
    "et-xmlfile==2.0.0",
    "eval-type-backport==0.2.2",
    "exceptiongroup==1.3.0",
    "executing==2.2.1",
    "fastapi==0.135.1",
    "filelock==3.25.2",
    "flatbuffers==25.2.10",
    "frozenlist==1.8.0",
    "fsspec==2025.9.0",
    "google-cloud-storage>=3.10.0",
    "googleapis-common-protos==1.73.0",
    "graphviz==0.21",
    "groq==0.34.1",
    "grpcio==1.78.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "huggingface-hub==0.36.2",
    "humanfriendly==10.0",
    "hume==0.13.10",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "iniconfig==2.1.0",
    "ipykernel==6.30.1",
    "ipython==8.37.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "langchain-core==1.2.19",
    "langchain-groq==1.1.2",
    "langgraph==1.1.2",
    "langgraph-checkpoint==4.0.1",
    "langgraph-checkpoint-postgres==3.0.4",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.11",
    "langsmith==0.7.17",
    "lightning-sdk==2025.11.13.post0",
    "livekit==1.1.2",
    "livekit-agents==1.4.4",
    "livekit-api==1.1.0",
    "livekit-blingfire==1.1.0",
    "livekit-plugins-deepgram==1.4.4",
    "livekit-plugins-groq==1.4.4",
    "livekit-plugins-openai==1.4.4",
    "livekit-plugins-silero==1.4.4",
    "livekit-plugins-turn-detector==1.4.4",
    "livekit-protocol==1.1.2",
    "locket==1.0.0",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "nest-asyncio==1.6.0",
    "netcdf4==1.7.2",
    "networkx==3.4.2",
    "numpy==2.2.6",
    "onnxruntime==1.23.2",
    "openai==2.26.0",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.39.1",
    "opentelemetry-exporter-otlp==1.39.1",
    "opentelemetry-exporter-otlp-proto-common==1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.39.1",
    "opentelemetry-exporter-otlp-proto-http==1.39.1",
    "opentelemetry-proto==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==25.0",
    "pandas==2.3.0",
    "parso==0.8.5",
    "partd==1.4.2",
    "pillow==12.1.1",
    "platformdirs==4.4.0",
    "plivo==4.59.3",
    "pluggy==1.6.0",
    "portalocker==3.2.0",
    "prometheus-client==0.24.1",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "psutil==7.1.0",
    "psycopg==3.3.3",
    "psycopg-binary==3.3.3",
    "psycopg-pool==3.3.0",
    "pure-eval==0.2.3",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pyreadline3==3.5.4",
    "pysat==3.2.2",
    "pytest==8.4.2",
    "pytest-asyncio==1.3.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.6",
    "pytz==2025.2",
    "pywin32==311",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "regex==2026.2.28",
    "requests==2.31.0",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "safetensors==0.7.0",
    "scipy==1.15.3",
    "shellingham==1.5.4",
    "simple-term-menu==1.6.6",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sounddevice==0.5.5",
    "stack-data==0.6.3",
    "starlette==0.52.1",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tokenizers==0.22.2",
    "tomli==2.2.1",
    "toolz==1.0.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.57.1",
    "twilio==9.8.5",
    "typer==0.24.1",
    "types-protobuf==4.25.0.20240417",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uuid-utils==0.14.1",
    "uv==0.8.3",
    "uvicorn==0.41.0",
    "watchfiles==1.1.1",
    "wcwidth==0.2.13",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "wget==3.2",
    "xarray==2025.6.1",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120
exclude = [".claude", "venv", ".venv"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["agents", "api", "db", "services"]

[tool.coverage.run]
omit = ["tests/*", ".claude/*", "venv/*"]

[tool.coverage.report]
show_missing = true
fail_under = 0
